requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=0.12.0=pypi_0
astor=0.8.1=pypi_0
backcall=0.2.0=py_0
beautifulsoup4=4.9.3=pyha847dfd_0
blas=1.0=mkl
brotlipy=0.7.0=py36h8f6f2f9_1001
bs4=4.9.3=hd3eb1b0_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.10.14=0
certifi=2020.6.20=py36_0
cffi=1.14.5=py36hc120d54_0
chardet=4.0.0=py36h5fab9bb_1
click=7.1.2=pyh9f0ad1d_0
cryptography=3.4.7=py36hb60f036_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py36_0
dataclasses=0.8=pyh4f3eec9_6
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
ethnicolr=0.5.0=pypi_0
expat=2.3.0=h2531618_2
ffmpeg=4.3=hf484d3e_0
filelock=3.0.12=pyh9f0ad1d_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
fuzzywuzzy=0.18.0=pyhd8ed1ab_0
gast=0.2.2=pypi_0
glib=2.68.0=h36276a3_0
gmp=6.2.1=h2531618_2
gnutls=3.6.5=h71b1129_1002
google-pasta=0.2.0=pypi_0
grpcio=1.37.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.9.0=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
intel-openmp=2020.2=254
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=py36_0
jedi=0.18.0=py36h06a4308_1
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py36_0
keras=2.2.4=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py36h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=21_mkl
libcblas=3.8.0=21_mkl
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h2828fa1_18
libgfortran-ng=7.3.0=hdf63c60_0
libgomp=9.3.0=h2828fa1_18
libiconv=1.15=h63c8f33_5
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=h6de172a_18
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=pypi_0
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=hbb512f6_0
ninja=1.10.2=py36hff7bd54_0
nltk=3.6.1=pyhd3eb1b0_0
numpy=1.16.4=pypi_0
olefile=0.46=py36_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h7f98852_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pyh44b312d_0
pandas=0.24.2=pypi_0
parso=0.8.0=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py36_0
pickleshare=0.7.5=py36_0
pillow=8.1.2=py36he98fc37_0
pip=21.0.1=py36h06a4308_0
prompt-toolkit=3.0.8=py_0
protobuf=3.15.8=pypi_0
ptyprocess=0.6.0=py36_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.1=py_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
pysocks=1.7.1=py36h5fab9bb_3
python=3.6.13=hdb3f193_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-levenshtein=0.12.2=py36h8f6f2f9_0
python_abi=3.6=1_cp36m
pytorch=1.8.1=py3.6_cuda10.2_cudnn7.6.5_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=pypi_0
pyzmq=19.0.2=py36he6710b0_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
regex=2021.3.17=py36h8f6f2f9_0
requests=2.25.1=pyhd3deb0d_0
sacremoses=0.0.43=pyh9f0ad1d_0
scikit-learn=0.23.2=py36hb6e6923_3
scipy=1.5.2=py36h0b6359f_0
setuptools=52.0.0=py36h06a4308_0
sip=4.19.8=py36hf484d3e_0
six=1.15.0=py36h06a4308_0
soupsieve=2.2.1=pyhd3eb1b0_0
sqlite=3.35.2=hdfb4753_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.1=py36h589faa6_0
torchaudio=0.8.1=py36
torchvision=0.9.1=py36_cu102
tornado=6.1=py36h27cfd23_0
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=4.3.3=py36_0
transformers=4.4.2=pyhd8ed1ab_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pyhd8ed1ab_0
wcwidth=0.2.5=py_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
